Refactor remaining package mocks
