+++ packages/leancode_lint/test/mock_libraries/bloc.dart
@@ -1,0 +1,13 @@
+part of '../mock_libraries.dart';
+
+mixin MockBloc on AnalysisRuleTest {
+  @override
+  void setUp() {
+    newPackage('bloc').addFile('lib/bloc.dart', '''
+abstract class Cubit<State> extends BlocBase<State> {
+  Cubit(State initialState) : super(initialState);
+}
+''');
+    super.setUp();
+  }
+}
