--- packages/leancode_lint/test/mock_utils.dart
@@ -1,9 +1,0 @@
-import 'package:analyzer_testing/analysis_rule/analysis_rule.dart';
-
-extension AnalysisRuleTestMockAppend on AnalysisRuleTest {
-  void appendFile(String path, String content) {
-    final file = getFile(path);
-    final oldContent = file.readAsStringSync();
-    file.writeAsStringSync('$oldContent\n$content');
-  }
-}
