+++ packages/leancode_lint/test/test_cases/add_cubit_suffix_for_cubits_test.dart
@@ -12,11 +12,11 @@ void main() {
 }
 
 @reflectiveTest
-class AddCubitSuffixForYourCubitsTest extends AnalysisRuleTest {
+class AddCubitSuffixForYourCubitsTest extends AnalysisRuleTest
+    with MockBloc, MockFlutterBloc {
   @override
   void setUp() {
     rule = AddCubitSuffixForYourCubits();
-    addMocks([.bloc, .flutterBloc]);
 
     super.setUp();
   }
